Add deprecation warning for TyrianApp
